fix(coverage): stop the poll-loop flake instead of re-running it
`Coverage (… / leviath-runtime)` has been failing intermittently on PRs that do
not touch the crate, most recently on #273 at `host.rs:2769`. Every time, all
1133 tests pass and the gate reports one uncovered region.

The region is the body of a test helper:

    while !done(&host.world_mut().lane_snapshot()) {
        tokio::time::sleep(Duration::from_millis(5)).await;   // <- 0-hit
    }

`await_lane` polls a background tool lane. When the lane happens to reach the
awaited state before the first check, the loop body never runs and the gate
reads a legitimately-unneeded sleep as a coverage hole.
`interaction_hub.rs`'s `wait_for_prompt` is the same shape, and its own comment
explains why the polling cannot be replaced by a fixed number of yields.

**Reordering the loop does not fix this, and I had already measured that** - it
moves the hole rather than closing it. That is inherent: whether a poll loop
needs to iterate is a fact about scheduling, so *some* edge of it is always
timing-dependent. There is no arrangement of a poll loop whose every region is
deterministically reachable.

So the fix is not to make the scaffolding coverable but to stop measuring it.
Both test modules move to sibling files, which is llvm-cov's own default
exclusion and the layout this repo already sanctions for exactly this case: a
test module whose *helpers* would otherwise need tests of their own. 24 files
already use it, and CONTRIBUTING now describes the rule.

Pure move - the modules stay children of the same parent, so `use super::*`
and every path inside resolve unchanged. 1133 tests, same count, all passing.

Verified the exclusion is the intended one and not wider: `host.rs` and
`interaction_hub.rs` are still measured and still 100% on lines, functions and
regions; `host_tests.rs` and `interaction_hub_tests.rs` appear nowhere in the
report. No production code left measurement.
